refactor: move useAuth hook to a separate file and update imports across components
